backend/requirements.txt
Django==5.2.1
psycopg==3.2.9
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
requests==2.32.3
firebase-admin==6.8.0